+++ src/lib.rs
@@ -75,7 +75,7 @@ pub use accept::Accept;
 pub use adam::minimize_adam;
 pub use control::Control;
 pub use error::{Error, Result};
-pub use highs_kind::{HighsCrossover, HighsSolverKind};
+pub use highs_kind::{HighsCCallback, HighsCallbackKind, HighsCrossover, HighsSolverKind};
 pub use fire::FireKind;
 pub use irc_trust::{sqrt_masses_3n, IrcTrust};
 pub use qn_irc::{